apps/api/requirements.txt
# Frame API — pinned minimums; see README / CONTEXT for optional keys (Hive, Meta, etc.)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
cryptography>=43.0.0
httpx>=0.27.0
python-multipart>=0.0.9
imagehash>=4.3.1
Pillow>=10.0.0
pytesseract>=0.3.10
anthropic>=0.25.0
yt-dlp>=2024.12.0
faster-whisper>=1.0.0
# Dossier / worker stack
asyncpg>=0.30.0
pgvector>=0.3.0
redis>=5.0.0
arq>=0.26.0
python-dotenv>=1.0.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
